Plot/pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plot-mcp"
version = "0.1.0"
description = "Plot MCP server implementation using Model Context Protocol"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=0.1.0",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.95.0",
    "uvicorn>=0.21.0",
    "pydantic>=1.10.0",
    "pytest>=7.2.0",
    "requests>=2.28.0",
    "pandas>=1.5.0",
    "matplotlib>=3.6.0",
    "seaborn>=0.12.0",
    "openpyxl>=3.0.0",
    "xlrd>=2.0.0",
    "fastmcp",
    "numpy<2.0.0"
]

[project.optional-dependencies]
test = ["pytest>=7.0.0", "httpx"]

[project.scripts]
plot-mcp = "plot.server:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["plot*"]

[tool.setuptools.package-dir]
"plot" = "src/plot"

[dependency-groups]
dev = [
    "pytest>=8.4.0",
]

[tool.uv.workspace]
members = [
    "plot-mcp",
]
